requirements.txt
python-socketio
pynput
pyfiglet
pycryptodome
pyfiglet
colorama